Archive ioslave: Better error handling in ArchiveProtocolBase::checkNewFile()
The resolution of the archive pathname did not distinguish between a
pathname component not existing and any other error (e.g. permission denied).
This resulted in just a generic error being shown.

If the archive is really a directory, ensure that the redirected URL
has a valid protocol to avoid an assert.