smb: no discovery cap on file listing
File and share listing can take a very long time as they are IO bound on
the remote, so if one is listing a large directory on a HDD it may well
take minutes. Equally sending huge listings over the wire may take a
while if not enough throughput is available. So don't cap the execution
of these discoveries at all.

Server discovery retains the timeout, mostly because I don't want to
remove it in 20.08 though an argument could be made that we should trust
in the reliability of the discovery stacks to not get stuck on random
unforseen misbehavior of servers.

BUG: 427644
FIXED-IN: 20.08.3