Filter out resource forks when filtering for images
The pages in a comic book are just a set of images, ordered by filename.
When there is a resource fork (as created by MacOS when using the
built-in ZIP support), the metadata files (which have the same suffix as
the corresponding image files) are sorted first.

For e.g a two page book the files `book_01.png`, `book_02.png` are sorted
after and hidden by `__MACOSX/._book_01.png`, `__MACOSX._book_02.png`.