nfs: Improve the file handle cache handling and lookup
Ensure that URLs are cleaned of trailing slashes before using them
as a path or key into the file handle cache, so that the cache
matching will not be affected.

Do not store an invalid file handle in the cache, otherwise it can
end up getting used later.

Verify that a NFSFileHandle is valid by looking at the data/link size,
rather than using the separate m_isInvalid flag.  This automatically
verifies the validity of NFS file handles returned from an RPC call.