nfs (v3): Allow listing of virtual directories even if they cannot be mounted
When attempting to mount the server's exported filesystems in
NFSProtocolV3::openConnection(), record the export path even if the
mount failed.  This allows them to be listed as virtual directories,
even though any other access needs a file handle.

Detect that situation in NFSProtocol::getFileHandle() and return an error.

Detect and warn about a mount which may have failed due to reserved port
restrictions.

Correct the virtual directory path prefix check in NFSProtocolV3::listDir().