nfs (v3): Centralise KIO error tracking and reporting
Instead of the NFSProtocol calling either SlaveBase::error() or
SlaveBase::finished(), record the error in NFSSlave::setError() and call
either error() or finished() in NFSSlave::finishOperation() as
appropriate.  Ensures that either of these are only called once even if
an error has to be reported by a low level function.