smb: do not assume rename files are different based on name
samba is transparently supporting case sensitivity/insensitivity based
on server capabilities so 'A' and 'a' may be the same file or not. To
that end when a rename operation would change the capitalization of a
file we need to do some extra work to figure out if that renaming would
constitute an overwrite or not. Specifically we'll need to stat the
source file and then compare the inode and device returned by libsmb to
figure out if they are the same file. If they are then this is an
in-place rename, not an overwrite and we'll skip over the error
conditions to do with the dst file already existing

BUG: 430585