nfs (v3): Fix resolution of invalid symlinks
That is, links that point nowhere or outside of the NFS server's
exported directories.  Avoid calling getFileHandle() on the target
of such a link, as it will use isValidPath() and set an error.
An error here is not appropriate for the result of the operation,
so just note the link as invalid.

Do not perform redundant (already done) checks in listDirCompat(),
and ensure that it returns a UDSEntry for ".".