nfs: Protocol version settings (general and host-specific)
This limits the maximum NFS protocol version that will be tried, and the
minimum protocol version that will be accepted.  It can be set in the
'kionfsrc' file either as a general setting, or overridden by
host-specific settings.

Also implement a setting for whether READDIRPLUS3 is to be used - some
servers don't seem to implement it properly.