[Thumbnail] Skip potentially slow thumb generation for symlinks
The calling code in kio already checks if the directory itself is on a
slow (e.g. network) filesystem, but following symlinks inside it subverts
this check.

Skipping thumbnails has the effect of prefering plain files instead of
symlinks. This still gives a good approximation of the directory contents
(and even a better one if the symlinks point to files in the same
directory, as duplicates are avoided).