Drop network: KIO slave
Never really left the stage of being a prototype for a hardware-centric
meta-protocol KIO slave, also not just file system centric, and
its presence is confusing compared to the remote KIO slave, which also
is tighter integrated with Dolpin & Co.

It had 12 years of chance* to take over, but did not, so time to take it
back from the playing field.

* https://frinring.wordpress.com/2009/04/20/network-kioslave-entered-kdereview/