kio_filenamesearch: fix crash due to KCoreDirLister changes
Error reporting was changed in KCoreDirlister, to match the original behaviour
we need to explicitly disable auto error handling.

This fixes a crash while searching in Dolphin, kio_filenamesearch intrenally
uses a dirlister, and a nested event loop, if there is an error while listing
some dir it tries to report it using a job ui delegate, but the parent job
could well be dead by that time, which causes a crash:
"QWidget: Cannot create a QWidget without QApplication"

Note that this crash only happens if KDE_FORK_SLAVES is _not_ set.

BUG: 438187