Remove configuration UI for JPEG thumbnailer
The options are very undiscoverable in dolphin and having
an option for this seems quite overkill to me. If people really want to
have this kind of configurability they should be using an image
viewer instead of looking at the thumbnails.

It also causes issues from a dependency kind of view.
Task: https://phabricator.kde.org/T14695