man: Call finished() instead of error() after a HTML error message
Konqueror with the WebEnginePart sees the KIO error and displays a
generic error, ignoring the formatted HTML explanation.  KHTML
puts up a deceptive blank error box.  Only WebKit ignores the error
and displays the HTML as intended.  Call finished() so that the HTML
will be displayed whatever viewing part is in use.

Partial revert of commit d5f2beb2, as noted in
https://invent.kde.org/network/konqueror/-/merge_requests/83