stop mtime duplication
applying the mtime was duplicated across multiple functions when indeed
it is largely the same code with only the actual setting being slightly
variable. use a template function with callback instead to share the
sharable bits.

also, since smbc_utime is used in two functions move that to its own
helper built on top to remove even the callback duplication