sftp: start splitting apart openconnectoin
it's way too large, way too complicated, way to repetitively concerned
with cleanup. fingerprinting is a fairly isolated bit, so split it into
a helper function. this also helps reduce the in-scope variables and in
turn makes it less likely that we forget to clean up something

also use scopeguards for cleanup in the new function. also to reduce
repetition and ensure cleanup