Fix id of krita thumbnailer not matching it's desktop name
Before KIO used KPluginMetaData for the thumbnailers, we would
use the DesktopEntryName for identifying the plugins.
Because the desktop files define no id, we would derive the id
from the library filename instead of the desktop name.

Consequently we should set the id to make sure the plugin can
be correctly identified.

BUG: 447507