kio_filenamesearch: don't use iterators when searching pending dirs
Looking at the backtrace from the bug report, it looks like the vector
could get reallocated when it changes capacity ...etc, which would
invalidate iterators.

Instead use a FIFO std::queue, since ideally the dirs should be searched in
the same order they're added to the pendingDirs container, and then
removed, and removing from the from of a vector is more costly than with a
queue (which uses a std::deque by default).

BUG: 448114
FIXED-IN: 22.04.3