mtp: support moving through get-put and increase concurrency to 2
this is a bit tricky. to support moving from one device to another
properly we need KIO to take control since doing it inside the worker is
exceptionally tricky with the current feature set. at the same time KIO
is a bit defective and presumes that it can have a concurrent getjob and
putjob running at the same time. to allow for that we now have 2 workers
instead of 1. technically they could be deadlocking each other,
practically they shouldn't because all the work is outsourced to the
kiod though anyway and thus getting multiplexed

BUG: 325574