mtp: Configure fails if smb is not being compiled
mtp/CMakeLists.txt uses cmake_{push,pop}_check_state() which need
CMakePushCheckState to be included.  This is done in smb/CMakeLists.txt,
but if libsmb is not available then this include will not have any
effect and the MTP configuration fails.

Since the include is needed in multiple places, move it to the
top level.