Thumbnailer: convert scaled images to sRGB
Converts preview to sRGB for proper viewing on most monitors (modern monitors should have a gamut equal to or greater than the sRGB color space). In this way the previews of HDR images (e.g. 32-bit per channel TIFF with linear color profile) are shown with the correct colors.

In the image on the left the result obtained with this MR, on the right the original result.

![image](/uploads/93dea51d51df6def81a4aa3284d124ab/image.png)