thumbnail: Fix heap-use-after-free in AudioCreator::create
There is a heap-use-after-free issue in `AudioCreator::create` resulting from storing the pointer to a temporary `QByteArray`'s data() in a pointer and accessing it after the byte-array has been freed (when the the temporary object was created on is over).

This fixes it by moving the `QByteArray` onto the stack, thus making it not temporary anymore, keeping it around until its data isn't needed anymore.

BUGS: 469458