[thumbnail] Limit bits per pixel to 32
PreviewJob allocates shared memory for transferring the thumbnail data. There it assumes max. 32 bits per pixel

Now when we load an image with 64 bpp we fail to send it because not enough memory was allocated

To avoid this convert the image to a 32bit format if neccesary

For a small thumbnail we don't really need that much color bits anyway

BUG: 484183