filenamesearch: Add missing translation domain
This kio-worker seems to have been always missing a translation domain.
Add the translation domain definition and Messages.sh, so the user
visible strings can get translatable.

Currently it has only one string ("Search Folder") which is the UDS
type it provides and can be seen on dolphin search info panel.

It might look small for a whole catalog, but we're doing the same
for other workers like `info`.