kcms/proxy,trash: follow HIG on capitalization
The HIG recommend sentence case for radiobutton, combobox, and text
field labels. The proxy and trash kcms currently are inconsistent with
regard to this and mix sentence and title case, sometimes within the
same group of controls.

This change makes the capitalization of these labels consistent with the
HIG recommendations.