kcms/webshortcuts: make labels more consistent
When adding a new entry or modifying an existing one, the terminology
switches from "Web search keywords" (used in the main dialog and e.g. in
the Plasma Search kcm, to "shortcut". The button for altering an
existing Web search keyword is labeled "Change", but the corresponding
dialog is titled "Modify [...]"; neither of them is the commonly used
term for editing entries in a listview, "Edit" (used, for example, in
the kwin rules, keyboard shortcut, or file association kcms). Most, but
not all instances of Web are capitalized. The add/modify dialog contains
the somewhat opaque term "charset". Finally, the HIG recommend
using a unicode ellipsis character "…" instead of three dots, but the
"New..." and "Change..." buttons still have individual dots.

This change modifies the New/Edit dialog labels to be consistent with
the ones in the main kcm (including not using the term "shortcut" and
matching button labels as well as column headers used there), labels the
button in the main kcm "Edit", capitalizes "Web" consistently (style
guides vary, but the HIG explicitly recommends capitalization for
"Internet" so capitalizing "Web" seems to be in the spirit of the
guidelines), de-abbreviates "Character set", and substitutes ellipsis
characters where appropriate.