filenamesearch: use `ripgrep-all` or `ripgrep` for content search if available
Use `ripgrep-all` (or `ripgrep` if the former is not installed) as the default
search backend for content search in the `kio-filenamesearch`, only fallback
to our own implementation if search in remote dirs, or neither of these tools are
available.

The command line arguments are not hardcoded. Instead, filenamesearch calls a
script to do the search. The default script is installed as
`/usr/share/kio_filenamesearch/kio-filenamesearch-grep`. The user can replace it
with `~/.local/share/kio_filenamesearch/kio-filenamesearch-grep`.