proxy: add a kded to detect WPAD problems
sentry revealed that some users enable WPAD and then have severe network
problems in (e.g.) discover to the point of producing
ApplicationNotResponding scenarios. help them find out about the
misconfiguration and guide them towards alternative proxy settings

this implements a very simple kded module that forks off a helper to try
and let libproxy do WPA discovery. if it doesn't manage to produce a
result within 1 second there's clearly something wrong and we'll throw a
notification at the user asking them to review their settings

additionally the settings now sport a contextualhelpbutton that gives
some context on what the auto discovery does and warning the user about
the problems of enabling it when not required