filenamesearch: Allow to search hidden files/folders
Previously we only searched visible files and this is distinct from
whether Dolphin shows hidden files or not. If you display hidden files
with a Ctrl-H and then search, you don't find any hidden files.

Add a "&includeHidden=yes" argument in the filenamesearch:/ url.

Add tests in the filenamesearch autotests for
searching for hidden files and within hidden folders.

CCBUG: 463830