kcms/webshortcuts: Make the "Preferred" table column more obvious
This switches the order of the columns in the table from "Name,
Shortcuts, Preferred" to "Name, Preferred, Shortcuts". This makes it
more obvious there even is a "Preferred" checkbox without having to
scroll the table.

This has the downside of eliding the keywords column, but the full list
is always available from the "Edit..." dialog. I think this is a
worthwhile tradeoff for removing the horizontal scrollbar.