Filenamesearch: Use QFlags to control the search scope.
Replace the isContent and includeHidden bools with QFlags.

A 'checkContent=yes' in the filenamesearch: URL now sets the
SearchContent flag. By default the SearchFileName is always set so
a content search now find matches in the filename and the contents
(of text/plain) files.

An 'includeHidden=yes' in the URL sets both IncludeHiddenFiles and
IncludeHiddenFolders flags, the default is not to search for hidden
files/folders.