Re-query available auth methods if result for pubkey auth was PARTIAL
Suppose the server uses pubkey authentication followed by a
keyboard-interactive authentication method (such as 2FA with Duo or
Google Authenticator). This is, roughly, what KIO does when trying to
authenticate with the server:

1. It calls `ssh_auth_list` from libssh to get the available
authentication methods. The return value will indicate that
`SSH_AUTH_METHOD_PUBLICKEY` is supported, but not
`SSH_AUTH_METHOD_INTERACTIVE` (because it is not *at this point in the
authentication process*).
2. Then it attempts to authenticate using a pubkey by calling
`ssh_userauth_pubkey_auto`. If the pubkey is accepted, the result is
`SSH_AUTH_PARTIAL` and not `SSH_AUTH_SUCCESS`. At this point,
`SSH_AUTH_METHOD_INTERACTIVE` *is* supported (and required).
3. Because the set of authentication methods was not queried again, KIO
still thinks keyboard-interactive authentication is not supported, and
therefore it does not try this authentication method. As a result, the
authentication process fails.

This commit fixes this issue by re-querying supported authentication
methods if the result from the pubkey authentication attempt is
`SSH_AUTH_PARTIAL`.