[thumbnail] Don't create a temporary QApplication to determine the platformtheme name
The current code creates a qapp to query the platform theme name and use that to
configure the actual qapp.

We do that since we want to use the offscreen platformtheme with our actual platformtheme
plugin.

However by instantiating the KDEPlatformTheme we trigger KWindowSystem's platform detection
code, which will cache the platform information (X11 or Wayland). Later we want that to be
'Unknown' though.

This trips up other code that relies on the KWindowSystem platform check before doing
windowsystem-specific stuff.

What we really want is a way to obtain the platformtheme name without creating an actual
instance.

Qt doesn't have API to do that, so copy the bit of code that determines the platform from
QGenericUnixServices

BUG: 504126