kio_filenamesearch: Add options to search for a phrase or regular expression.
Look for "syntax=phrase" or "syntax=regex" options, if a searching for a phrase,
escape the regex special characters.

A search for a phrase still allows a flexible whitespace match, rather than
doing a literal match, which should exactly match spaces.

Extend the filenamesearchtests and filenamesearchexternaltests to include phrase
searches.