wpad-detector: delay setup by 60s
otherwise we may erronously raise the warning if the network is just
slow to come up. something that may in particular happen if VPNs are
involved, which is all the more likely considering legitimate wpad use
cases are more on the corporate side.

with that in mind the reachabilityChanged also only triggers checks
after the initial delay because the network may be up, but not the VPN.

bit awkward this