sftp: Keep permissions of copied file from remote
When copying from the remote to local, the initial copy unconditionally
adds owner_write permission for some unspecified NFS issue.

Make sure that after copying the file, the permissions on the destination
file become identical to that of the source file.

Copying from local to remote is not affected and is already correct.